products/tier1/bot-telegram/src/requirements.txt
# Herda do requirements.txt principal
../../requirements.txt

# Específico do bot
python-telegram-bot==20.7
langchain==0.1.0
langchain-google-genai==0.0.5
langchain-community==0.0.13
faiss-cpu==1.7.4
openai==1.3.9
google-generativeai==0.3.0
psutil>=5.9.0  # Para monitoramento de memória

